requirements.txt
streamlit
groq
PyPDF2 
Pillow 
google
googlesearch-python